🚚 Refactor component structure
